+++ app/models/dossier_eleve.rb
@@ -154,11 +154,14 @@ class DossierEleve < ActiveRecord::Base
 
   def options_triees_par_rang
     options_par_rang = {}
+    options_facultatives = []
     options_pedagogiques.each do |option|
       rang_option = option.mef_options_pedagogiques.find_by(mef: mef_destination)&.rang_option
-      options_par_rang.merge!(rang_option => option)
+      options_par_rang[rang_option] = option if rang_option.present?
+      options_facultatives << option unless rang_option.present?
     end
-    Hash[options_par_rang.sort].map { |_rang, option| option }
+    options_obligatoires = Hash[options_par_rang.sort].map { |_rang, option| option }
+    options_obligatoires + options_facultatives
   end
 
 end